Update scale-deployment.ps1
This change will allow users to add VMs to an existing deployment.  The user would be able to call the script again but set starting suffix as the next VM that they want and add any number of additional VMs. Current logic doesn't allow for this use case.

For e.g. I have vm0, vm1, and vm2 deployed but want to add vm3 and vm4 therefore startingsuffix = 3 and count = 1.  This now works.